split up partnerIntegrations.ts
