+++ src/background/auth/partnerIntegrations/types.ts
@@ -1,0 +1,31 @@
+/*
+ * Copyright (C) 2024 PixieBrix, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+/**
+ * A principal on a remote service, e.g., an Automation Anywhere Control Room.
+ */
+export type PartnerPrincipal = {
+  /**
+   * The hostname of the remote service, e.g., the Automation Anywhere Control Room.
+   */
+  hostname: string;
+
+  /**
+   * The principal unique id, or null for OAuth-based integrations.
+   */
+  principalId: string | null;
+};
