+++ src/tsconfig.strictNullChecks.json
@@ -136,8 +136,8 @@
     "./background/messenger/registration.ts",
     "./background/navigation.ts",
     "./background/partnerHandlers.ts",
-    "./background/partnerIntegrations.test.ts",
-    "./background/partnerIntegrations.ts",
+    "./background/auth/partnerIntegrations/partnerIntegrations.test.ts",
+    "./background/auth/partnerIntegrations/partnerIntegrations.ts",
     "./background/performConfiguredRequest.test.ts",
     "./background/proxyUtils.test.ts",
     "./background/proxyUtils.ts",
